requirements.txt
setuptools
wheel
python-vlc
numpy
vosk
psutil
pyaudio
rapidfuzz
scikit-learn
dateparser
flask
paramiko
cryptography
flask-socketio
eventlet
face_recognition
opencv-python
llama-cpp-python
huggingface-hub
pychromecast
paho-mqtt
faster-whisper
piper-tts
padatious
fann2
speedtest-cli
speedtest-cli
